feat add @JSONType annotation and remove outdated processor
