+++ codegen/src/main/java/com/alibaba/fastjson2/internal/processor/JSONCompiledAnnotationProcessor.java
@@ -1,14 +1,23 @@
 package com.alibaba.fastjson2.internal.processor;
 
-import com.alibaba.fastjson2.JSONException;
-import com.alibaba.fastjson2.JSONPath;
 import com.alibaba.fastjson2.JSONReader;
+import com.alibaba.fastjson2.annotation.JSONCompiled;
 import com.alibaba.fastjson2.internal.asm.Label;
-import com.alibaba.fastjson2.internal.codegen.Block;
-import com.alibaba.fastjson2.internal.codegen.ClassWriter;
-import com.alibaba.fastjson2.internal.codegen.MethodWriter;
-import com.alibaba.fastjson2.internal.codegen.Opcodes;
-import com.alibaba.fastjson2.reader.*;
+import com.alibaba.fastjson2.reader.FieldReader;
+import com.alibaba.fastjson2.reader.ObjectReader;
+import com.alibaba.fastjson2.reader.ObjectReaderAdapter;
+import com.alibaba.fastjson2.reader.ObjectReaders;
+import com.sun.tools.javac.api.JavacTrees;
+import com.sun.tools.javac.code.Flags;
+import com.sun.tools.javac.code.TypeTag;
+import com.sun.tools.javac.processing.JavacProcessingEnvironment;
+import com.sun.tools.javac.tree.JCTree;
+import com.sun.tools.javac.tree.TreeMaker;
+import com.sun.tools.javac.tree.TreeTranslator;
+import com.sun.tools.javac.util.Context;
+import com.sun.tools.javac.util.List;
+import com.sun.tools.javac.util.ListBuffer;
+import com.sun.tools.javac.util.Names;
 
 import javax.annotation.processing.*;
 import javax.lang.model.SourceVersion;
@@ -19,15 +28,13 @@ import javax.tools.JavaFileObject;
 
 import java.io.IOException;
 import java.io.Writer;
-import java.lang.reflect.Modifier;
 import java.lang.reflect.Type;
 import java.nio.charset.StandardCharsets;
 import java.util.*;
+import java.util.stream.Collectors;
 
-import static com.alibaba.fastjson2.internal.codegen.Opcodes.*;
-import static com.alibaba.fastjson2.internal.processor.CodeGenUtils.fieldObjectReader;
-import static com.alibaba.fastjson2.internal.processor.CodeGenUtils.fieldReader;
-import static com.alibaba.fastjson2.internal.processor.JavacTreeUtils.unwrapProcessingEnv;
+import static com.alibaba.fastjson2.internal.processor.CodeGenUtils.*;
+import static com.alibaba.fastjson2.internal.processor.JavacTreeUtils.*;
 import static com.alibaba.fastjson2.util.JDKUtils.ARRAY_BYTE_BASE_OFFSET;
 import static com.alibaba.fastjson2.util.JDKUtils.UNSAFE;
 
@@ -41,222 +48,184 @@ import static com.alibaba.fastjson2.util.JDKUtils.UNSAFE;
 })
 public class JSONCompiledAnnotationProcessor
         extends AbstractProcessor {
+    private Messager messager;
+    private JavacTrees javacTrees;
+    private Names names;
+
     @Override
     public synchronized void init(ProcessingEnvironment processingEnv) {
-        super.init(unwrapProcessingEnv(processingEnv));
+        ProcessingEnvironment unwrappedProcessingEnv = unwrapProcessingEnv(processingEnv);
+        super.init(unwrappedProcessingEnv);
+        this.messager = unwrappedProcessingEnv.getMessager();
+        this.javacTrees = JavacTrees.instance(unwrappedProcessingEnv);
+        Context context = ((JavacProcessingEnvironment) unwrappedProcessingEnv).getContext();
+        this.names = Names.instance(context);
+        initialize(TreeMaker.instance(context), names, unwrappedProcessingEnv.getElementUtils());
     }
 
     @Override
     public boolean process(Set<? extends TypeElement> annotations, RoundEnvironment roundEnv) {
-        if (roundEnv.processingOver() || annotations.isEmpty()) {
-            return false;
-        }
-
         Analysis analysis = new Analysis(processingEnv);
         Set<? extends Element> compiledJsons = roundEnv.getElementsAnnotatedWith(analysis.jsonCompiledElement);
-
         if (!compiledJsons.isEmpty()) {
             analysis.processAnnotation(analysis.compiledJsonType, compiledJsons);
         }
 
         Map<String, StructInfo> structs = analysis.analyze();
-        final Map<String, StructInfo> generatedFiles = new HashMap<>();
-        final List<Element> originatingElements = new ArrayList<>();
-//        Set<? extends Element> jsonConverters = roundEnv.getElementsAnnotatedWith(analysis.converterElement);
-//        Map<String, Element> configurations = analysis.processConverters(jsonConverters);
+        Set<? extends Element> elementsAnnotatedWith = roundEnv.getElementsAnnotatedWith(JSONCompiled.class);
+        elementsAnnotatedWith.forEach(element -> {
+            StructInfo structInfo = structs.get(element.toString());
+            java.util.List<AttributeInfo> attributeInfos = structInfo.getReaderAttributes();
+            int fieldsSize = attributeInfos.size();
+            Class superClass = getSuperClass(attributeInfos.size());
 
-        for (Map.Entry<String, StructInfo> entry : structs.entrySet()) {
-            String typeName = entry.getKey();
-            StructInfo info = entry.getValue();
+            JCTree tree = javacTrees.getTree(element);
+            pos(tree.pos);
+            tree.accept(new TreeTranslator() {
+                @Override
+                public void visitClassDef(JCTree.JCClassDecl beanClassDecl) {
+                    super.visitClassDef(beanClassDecl);
+                    String beanFQN = beanClassDecl.sym.toString();
+                    if (element.toString().equals(beanFQN)) {
+                        // initialization
+                        String innerClassFQN = findConverterName(structInfo);
+                        int dotIdx = innerClassFQN.lastIndexOf('.');
+                        String innerClassName = innerClassFQN.substring(dotIdx + 1);
+                        JCTree.JCExpression beanType = dotIdx == -1 ? ident(beanFQN) : qualIdent(beanFQN);
+                        JCTree.JCNewClass beanNew = newClass(null, null, beanType, null, null);
+                        JCTree.JCIdent objectType = ident("Object");
+                        // add annotation
+                        addAnnotation(beanClassDecl);
 
-            String classNamePath = findConverterName(info);
-            try {
-                JavaFileObject converterFile = processingEnv.getFiler().createSourceFile(classNamePath, info.element);
-                try (Writer writer = converterFile.openWriter()) {
-                    buildCode(writer, processingEnv, entry.getKey(), info, structs);
-                    generatedFiles.put(classNamePath, info);
-                    originatingElements.add(info.element);
-                } catch (IOException e) {
-                    processingEnv.getMessager().printMessage(Diagnostic.Kind.ERROR,
-                            "Failed saving compiled json serialization file " + classNamePath);
-                }
-            } catch (IOException e) {
-                processingEnv.getMessager().printMessage(Diagnostic.Kind.ERROR,
-                        "Failed creating compiled json serialization file " + classNamePath);
-            }
-        }
+                        // generate inner class
+                        JCTree.JCClassDecl innerClass = genInnerClass(innerClassName, superClass);
 
-        return false;
-    }
+                        // generate fields if necessary
+                        final boolean generatedFields = fieldsSize < 128;
+                        if (generatedFields) {
+                            innerClass.defs = innerClass.defs.prependList(genFields(attributeInfos, superClass));
+                        }
 
-    static String findConverterName(StructInfo structInfo) {
-        int dotIndex = structInfo.binaryName.lastIndexOf('.');
-        String className = structInfo.binaryName.substring(dotIndex + 1);
-        if (dotIndex == -1) {
-            return className + "_FASTJSONReader";
-        }
-        String packageName = structInfo.binaryName.substring(0, dotIndex);
-        return packageName + '.' + className + "_FASTJSONReader";
-    }
+                        // generate constructor
+                        innerClass.defs = innerClass.defs.append(genConstructor(beanType, beanNew, attributeInfos, superClass, generatedFields));
 
-    private static void buildCode(
-            final Writer code,
-            final ProcessingEnvironment environment,
-            final String className,
-            final StructInfo si,
-            final Map<String, StructInfo> structs
-    ) throws IOException {
-        final String generateFullClassName = findConverterName(si);
-        final int dotIndex = generateFullClassName.lastIndexOf('.');
-        final String generateClassName = generateFullClassName.substring(dotIndex + 1);
+                        // generate createInstance
+                        innerClass.defs = innerClass.defs.append(genCreateInstance(objectType, beanNew));
 
-        String packageName = null;
-        if (dotIndex != -1) {
-            packageName = generateFullClassName.substring(0, dotIndex);
-        }
+                        // generate readObject
+                        innerClass.defs = innerClass.defs.append(genReadObject(objectType, beanType, beanNew, attributeInfos, structInfo, false));
 
-        List<AttributeInfo> fields = si.getReaderAttributes();
+                        // link with inner class
+                        beanClassDecl.defs = beanClassDecl.defs.append(innerClass);
 
-        Class supperClass = CodeGenUtils.getSuperClass(fields.size());
-        ClassWriter cw = new ClassWriter(packageName, generateClassName, supperClass, new Class[0]);
-
-        final boolean generatedFields = fields.size() < 128;
-        if (generatedFields) {
-            genFields(fields, cw, supperClass);
-        }
-
-        {
-            MethodWriter mw = cw.method(
-                    Modifier.PUBLIC,
-                    "<init>",
-                    void.class,
-                    new Class[]{},
-                    new String[]{}
-            );
-
-            Op[] fieldReaders = new Op[fields.size()];
-            for (int i = 0; i < fields.size(); i++) {
-                AttributeInfo attr = fields.get(i);
-                String fieldType = attr.type.toString();
-                boolean fieldTypeIsClass = fieldType.indexOf('<') == -1;
-                Op fieldReader;
-                if (attr.setMethod != null) {
-                    String methodName = attr.setMethod.getSimpleName().toString();
-                    fieldReader = invokeStatic(
-                            ObjectReaders.class,
-                            "fieldReaderWithMethod",
-                            ldc(attr.name),
-                            getStatic(className, "class"),
-                            ldc(methodName)
-                    );
-                } else if (attr.field != null) {
-                    String fieldName = attr.field.getSimpleName().toString();
-                    if (fieldName.equals(attr.name)) {
-                        fieldReader = invokeStatic(
-                                ObjectReaders.class,
-                                "fieldReaderWithField",
-                                ldc(fieldName),
-                                getStatic(className, "class")
-                        );
-                    } else {
-                        fieldReader = invokeStatic(
-                                ObjectReaders.class,
-                                "fieldReaderWithField",
-                                ldc(attr.name),
-                                getStatic(className, "class"),
-                                ldc(fieldName)
-                        );
+                        // generate source file
+                        genSource(innerClassFQN, structInfo, innerClass);
                     }
-
-                    fieldReaders[i] = fieldReader;
-                } else {
-                    throw new IOException("TODO");
                 }
-                fieldReaders[i] = fieldReader;
-            }
-            mw.invoke(SUPER,
-                    "<init>",
-                    getStatic(className, "class"),
-                    Opcodes.ldc(si.typeKey),
-                    Opcodes.ldc(null),
-                    Opcodes.ldc(si.readerFeatures),
-                    Opcodes.methodRef(className, "new"),
-                    Opcodes.ldc(null),
-                    Opcodes.allocateArray(FieldReader.class, null, fieldReaders)
-            );
-
-            genInitFields(fields.size(), generatedFields, "fieldReaders", mw, supperClass);
-        }
-
-        {
-            MethodWriter mw = cw.method(
-                    Modifier.PUBLIC,
-                    "createInstance",
-                    Object.class,
-                    new Class[]{long.class},
-                    new String[]{"features"}
-            );
-
-            mw.ret(
-                    allocate(className)
-            );
-        }
-
-        genMethodReadObject(className, si, fields, cw, false);
-
-        code.write(cw.toString());
+            });
+        });
+        return true;
     }
 
-    static void genInitFields(
-            int fieldReaderArray,
-            boolean generatedFields,
-            String fieldReaders,
-            MethodWriter mw,
-            Class objectReaderSuper
-    ) {
-        if (objectReaderSuper != ObjectReaderAdapter.class || !generatedFields) {
-            return;
-        }
-
-        for (int i = 0; i < fieldReaderArray; i++) {
-            mw.putField(fieldReader(i), Opcodes.arrayGet(var(fieldReaders), Opcodes.ldc(i)));
+    private void addAnnotation(JCTree.JCClassDecl beanClassDecl) {
+        List<JCTree.JCAnnotation> annotations = beanClassDecl.mods.annotations;
+        Set<String> annotationTypes = annotations.stream().map(a -> a.getAnnotationType().type.tsym.toString()).collect(Collectors.toSet());
+        JCTree.JCExpression jsonTypeIdent = qualIdent("com.alibaba.fastjson2.annotation.JSONType");
+        if (!annotationTypes.contains(jsonTypeIdent.type.tsym.toString())) {
+            JCTree.JCAnnotation annotation = annotation(jsonTypeIdent, null);
+            beanClassDecl.mods.annotations = annotations.prepend(annotation);
         }
     }
 
-    static void genFields(List<AttributeInfo> fields, ClassWriter cw, Class objectReaderSuper) {
-        int fieldCount = fields.size();
-        if (objectReaderSuper == ObjectReaderAdapter.class) {
-            for (int i = 0; i < fieldCount; i++) {
-                cw.field(Modifier.PUBLIC, fieldReader(i), FieldReader.class);
+    private JCTree.JCClassDecl genInnerClass(String className, Class superClass) {
+        JCTree.JCClassDecl innerClass = defClass(Flags.PUBLIC | Flags.STATIC,
+                className,
+                null,
+                qualIdent(superClass.getName()),
+                null,
+                null);
+        return innerClass;
+    }
+
+    private List<JCTree> genFields(java.util.List<AttributeInfo> attributeInfos, Class superClass) {
+        ListBuffer<JCTree> stmts = new ListBuffer<>();
+        int fieldsSize = attributeInfos.size();
+        JCTree.JCExpression fieldReaderType = qualIdent(FieldReader.class.getName());
+        JCTree.JCExpression objectReaderType = qualIdent(ObjectReader.class.getName());
+        if (superClass == ObjectReaderAdapter.class) {
+            for (int i = 0; i < fieldsSize; ++i) {
+                stmts.append(defVar(Flags.PRIVATE, fieldReader(i), fieldReaderType));
             }
 
-            for (int i = 0; i < fieldCount; i++) {
-                cw.field(Modifier.PUBLIC, fieldObjectReader(i), ObjectReader.class);
+            for (int i = 0; i < fieldsSize; i++) {
+                stmts.append(defVar(Flags.PRIVATE, fieldObjectReader(i), objectReaderType));
             }
         }
 
-        for (int i = 0; i < fields.size(); i++) {
-            AttributeInfo field = fields.get(i);
+        for (int i = 0; i < fieldsSize; ++i) {
+            AttributeInfo field = attributeInfos.get(i);
             String fieldType = field.type.toString();
-
-            if (fieldType.startsWith("java.util.List<")
-                    || fieldType.startsWith("java.util.Map<java.lang.String,")
-            ) {
-                cw.field(Modifier.PUBLIC, CodeGenUtils.fieldItemObjectReader(i), ObjectReader.class);
+            if (fieldType.startsWith("java.util.List<") || fieldType.startsWith("java.util.Map<java.lang.String,")) {
+                stmts.append(defVar(Flags.PRIVATE, fieldItemObjectReader(i), objectReaderType));
             }
         }
+        return stmts.toList();
     }
 
-    public static void genMethodReadObject(
-            String className,
-            StructInfo si,
-            List<AttributeInfo> fields,
-            ClassWriter cw,
-            boolean jsonb
-    ) {
+    private JCTree.JCMethodDecl genConstructor(JCTree.JCExpression beanType, JCTree.JCNewClass beanNew, java.util.List<AttributeInfo> fields, Class superClass, boolean generatedFields) {
+        JCTree.JCLambda lambda = lambda(List.nil(), beanNew);
+        JCTree.JCExpression fieldReaderType = qualIdent(FieldReader.class.getName());
+        ListBuffer fieldReaders = new ListBuffer<>();
+        JCTree.JCExpression objectReadersType = qualIdent(ObjectReaders.class.getName());
+        int fieldsSize = fields.size();
+        for (int i = 0; i < fieldsSize; ++i) {
+            JCTree.JCMethodInvocation readerMethod = null;
+            AttributeInfo attributeInfo = fields.get(i);
+            if (attributeInfo.setMethod != null) {
+                String methodName = attributeInfo.setMethod.getSimpleName().toString();
+                readerMethod = method(field(objectReadersType, "fieldReaderWithMethod"), List.of(literal(attributeInfo.name), field(beanType, names._class), literal(methodName)));
+            } else if (attributeInfo.field != null) {
+                String fieldName = attributeInfo.field.getSimpleName().toString();
+                if (fieldName.equals(attributeInfo.name)) {
+                    readerMethod = method(field(objectReadersType, "fieldReaderWithField"), List.of(literal(fieldName), field(beanType, names._class)));
+                } else {
+                    readerMethod = method(field(objectReadersType, "fieldReaderWithField"), List.of(literal(attributeInfo.name), field(beanType, names._class), literal(fieldName)));
+                }
+            } else {
+                messager.printMessage(Diagnostic.Kind.WARNING, "not implemented yet");
+            }
+            if (readerMethod != null) {
+                fieldReaders.append(readerMethod);
+            }
+        }
+        JCTree.JCNewArray fieldReadersArray = newArray(fieldReaderType, null, List.from(fieldReaders));
+        JCTree.JCMethodInvocation superMethod = method(ident(names._super), List.of(field(beanType, names._class), defNull(), defNull(), literal(TypeTag.LONG, 0L), defNull(), lambda, defNull(), fieldReadersArray));
+        ListBuffer<JCTree.JCStatement> stmts = new ListBuffer<>();
+        stmts.append(exec(superMethod));
+        // initialize fields if necessary
+        if (superClass == ObjectReaderAdapter.class && generatedFields) {
+            stmts.appendList(genInitFields(fieldsSize));
+        }
+        return defMethod(Flags.PUBLIC, names.init, type(TypeTag.VOID), null, null, null, block(stmts.toList()), null);
+    }
+
+    private List<JCTree.JCStatement> genInitFields(int fieldsSize) {
+        ListBuffer<JCTree.JCStatement> stmts = new ListBuffer<>();
+        for (int i = 0; i < fieldsSize; ++i) {
+            stmts.append(exec(assign(ident(fieldReader(i)), indexed(field(ident(names._this), "fieldReaders"), literal(i)))));
+        }
+        return stmts.toList();
+    }
+
+    private JCTree.JCMethodDecl genCreateInstance(JCTree.JCIdent objectType, JCTree.JCNewClass beanNew) {
+        JCTree.JCVariableDecl featuresVar = defVar(Flags.PARAMETER, "features", type(TypeTag.LONG));
+        return defMethod(Flags.PUBLIC, "createInstance", objectType, null, List.of(featuresVar), null, block(defReturn(beanNew)), null);
+    }
+
+    private JCTree.JCMethodDecl genReadObject(JCTree.JCIdent objectType, JCTree.JCExpression beanType, JCTree.JCNewClass beanNew, java.util.List<AttributeInfo> attributeInfos, StructInfo structInfo, boolean isJsonb) {
         int fieldNameLengthMin = 0, fieldNameLengthMax = 0;
-        for (int i = 0; i < fields.size(); ++i) {
-            String fieldName = fields.get(i).name;
+        for (int i = 0; i < attributeInfos.size(); ++i) {
+            String fieldName = attributeInfos.get(i).name;
 
             int fieldNameLength = fieldName.getBytes(StandardCharsets.UTF_8).length;
             if (i == 0) {
@@ -268,140 +237,198 @@ public class JSONCompiledAnnotationProcessor
             }
         }
 
-        MethodWriter mw = cw.method(
-                Modifier.PUBLIC,
-                "readObject",
-                Object.class,
-                new Class[]{JSONReader.class, Type.class, Object.class, long.class},
-                new String[]{"jsonReader", "fieldType", "fieldName", "features"}
-        );
+        JCTree.JCVariableDecl jsonReaderVar = defVar(Flags.PARAMETER, "jsonReader", qualIdent(JSONReader.class.getName()));
+        JCTree.JCIdent jsonReaderIdent = ident(jsonReaderVar.name);
+        JCTree.JCVariableDecl fieldTypeVar = defVar(Flags.PARAMETER, "fieldType", qualIdent(Type.class.getName()));
+        JCTree.JCVariableDecl fieldNameVar = defVar(Flags.PARAMETER, "fieldName", objectType);
+        JCTree.JCVariableDecl featuresVar = defVar(Flags.PARAMETER, "features", type(TypeTag.LONG));
 
-        Op jsonReader = var("jsonReader");
-        Op features = var("features");
-        Op fieldType = var("fieldType");
-        Op fieldName = var("fieldName");
-        OpName object = var("object");
+        JCTree.JCReturn nullReturn = defReturn(defNull());
 
-        mw.ifStmt(invoke(jsonReader, "nextIfNull"))
-                .ret(Opcodes.ldc(null));
+        ListBuffer<JCTree.JCStatement> readObjectBody = new ListBuffer<>();
 
-        mw.newLine();
-        mw.stmt(invoke(jsonReader, "nextIfObjectStart"));
+        JCTree.JCMethodInvocation nextIfNullMethod = method(field(jsonReaderIdent, "nextIfNull"));
+        readObjectBody.append(defIf(nextIfNullMethod, block(nullReturn), null));
 
-        OpName features2 = var("features2");
-        mw.declare(long.class, features2, bitOr(var("features"), getField(THIS, "features")));
-        mw.newLine();
+        readObjectBody.append(exec(method(field(jsonReaderIdent, "nextIfObjectStart"))));
 
-        mw.declare(className, object, allocate(className));
-        mw.newLine();
+        JCTree.JCVariableDecl features2Var = defVar(Flags.PARAMETER, "features2", type(TypeTag.LONG), binary(JCTree.Tag.BITOR, ident("features"), field(ident(names._this), "features")));
+        readObjectBody.append(features2Var);
 
-        String forLabel = null;
-        if (fields.size() > 6) {
-            forLabel = "_while";
-            mw.label(forLabel);
-        }
+        JCTree.JCVariableDecl objectVar = defVar(Flags.PARAMETER, "object", beanType, beanNew);
+        JCTree.JCIdent objectIdent = ident(objectVar.name);
+        readObjectBody.append(objectVar);
 
-        Block.WhileStmt whileStmt = mw.whileStmtStmt(
-                not(invoke(jsonReader, "nextIfObjectEnd"))
-        );
+        int fieldsSize = attributeInfos.size();
+
+        JCTree.JCLabeledStatement loopLabel = label("_while", null);
+        JCTree.JCWhileLoop loopHead = whileLoop(unary(JCTree.Tag.NOT, method(field(jsonReaderIdent, "nextIfObjectEnd"))), null);
+        ListBuffer<JCTree.JCStatement> loopBody = new ListBuffer<>();
 
         if (fieldNameLengthMin >= 2 && fieldNameLengthMax <= 43) {
-            genRead243(si, whileStmt, forLabel, fields, jsonReader, object);
+            loopBody.appendList(genRead243(attributeInfos, jsonReaderIdent, structInfo, loopLabel, objectIdent, false));
         }
 
-        OpName hashCode64 = var("hashCode64");
-        whileStmt.declare(long.class, hashCode64, invoke(jsonReader, "readFieldNameHashCode"));
-        whileStmt.newLine();
-
-        if (fields.size() <= 6) {
-            for (int i = 0; i < fields.size(); ++i) {
-                AttributeInfo field = fields.get(i);
-                long fieldNameHash = field.nameHashCode;
-                Block.IfStmt ifStmt = whileStmt.ifStmt(eq(hashCode64, ldc(fieldNameHash)));
-                genReadFieldValue(ifStmt, i, si, field, jsonReader, object, forLabel, jsonb);
-                ifStmt.continueStmt();
-                whileStmt.newLine();
+        JCTree.JCFieldAccess readFieldNameHashCode = field(jsonReaderIdent, "readFieldNameHashCode");
+        JCTree.JCVariableDecl hashCode64Var = defVar(Flags.PARAMETER, "hashCode64", type(TypeTag.LONG), cast(type(TypeTag.LONG), method(readFieldNameHashCode)));
+        JCTree.JCExpression hashCode64 = ident(hashCode64Var.name);
+        loopBody.append(hashCode64Var);
+        if (fieldsSize <= 6) {
+            for (int i = 0; i < fieldsSize; ++i) {
+                AttributeInfo attributeInfo = attributeInfos.get(i);
+                List<JCTree.JCStatement> readFieldValueStmts = genReadFieldValue(attributeInfo, jsonReaderIdent, i, structInfo, loopLabel, objectIdent, isJsonb);
+                loopBody.appendList(List.of(defIf(binary(JCTree.Tag.EQ, literal(TypeTag.LONG, attributeInfo.nameHashCode), hashCode64), block(readFieldValueStmts), null)));
             }
         } else {
-            Map<Integer, List<Long>> map = new TreeMap();
+            Map<Integer, java.util.List<Long>> map = new TreeMap();
             Map<Long, AttributeInfo> mapping = new TreeMap();
             Map<Long, Integer> mappingIndex = new TreeMap();
-
-            for (int i = 0; i < fields.size(); i++) {
-                AttributeInfo field = fields.get(i);
-                long fieldNameHash = field.nameHashCode;
+            for (int i = 0; i < fieldsSize; ++i) {
+                AttributeInfo attr = attributeInfos.get(i);
+                long fieldNameHash = attr.nameHashCode;
                 int hashCode32 = (int) (fieldNameHash ^ (fieldNameHash >>> 32));
-                List<Long> hashCode64List = map.computeIfAbsent(hashCode32, k -> new ArrayList<>());
+                java.util.List<Long> hashCode64List = map.computeIfAbsent(hashCode32, k -> new ArrayList<>());
                 hashCode64List.add(fieldNameHash);
-                mapping.put(fieldNameHash, field);
+                mapping.put(fieldNameHash, attr);
                 mappingIndex.put(fieldNameHash, i);
             }
-
             int[] hashCode32Keys = new int[map.size()];
-            {
-                int off = 0;
-                for (Integer key : map.keySet()) {
-                    hashCode32Keys[off++] = key;
-                }
+            int off = 0;
+            for (Integer key : map.keySet()) {
+                hashCode32Keys[off++] = key;
             }
             Arrays.sort(hashCode32Keys);
+            JCTree.JCVariableDecl hashCode32Var = getHashCode32Var(hashCode64);
+            loopBody.append(hashCode32Var);
+            JCTree.JCExpression hashCode32 = ident(hashCode32Var.name);
 
-            // int hashCode32 = (int)(hashCode64 ^ (hashCode64 >>> 32));
-            OpName hashCode32 = var("hashCode32");
-            whileStmt.declare(int.class, hashCode32, cast(eor(hashCode64, urs(hashCode64, ldc(32))), int.class));
-            Block.SwitchStmt switchStmt = whileStmt.switchStmt(hashCode32, hashCode32Keys);
-            for (int i = 0; i < switchStmt.labels(); i++) {
-                Block label = switchStmt.label(i);
-                List<Long> hashCode64Array = map.get(switchStmt.labelKey(i));
-
+            JCTree.JCLabeledStatement switchLabel = label("_switch", null);
+            ListBuffer<JCTree.JCCase> cases = new ListBuffer<>();
+            for (int i = 0; i < hashCode32Keys.length; ++i) {
+                java.util.List<Long> hashCode64Array = map.get(hashCode32Keys[i]);
+                List<JCTree.JCStatement> stmts = List.nil();
+                Long fieldNameHash = null;
                 if (hashCode64Array.size() == 1 && hashCode64Array.get(0) == hashCode32Keys[i]) {
-                    Long fieldNameHash = hashCode64Array.get(0);
+                    fieldNameHash = hashCode64Array.get(0);
                     int index = mappingIndex.get(fieldNameHash);
-                    AttributeInfo field = mapping.get(fieldNameHash);
-                    genReadFieldValue(label, index, si, field, jsonReader, object, forLabel, jsonb);
-                    label.continueStmt(forLabel);
+                    AttributeInfo attributeInfo = mapping.get(fieldNameHash);
+                    stmts = stmts.appendList(genReadFieldValue(attributeInfo, jsonReaderIdent, index, structInfo, loopLabel, objectIdent, isJsonb));
+                    stmts.append(defContinue(loopLabel));
                 } else {
                     for (int j = 0; j < hashCode64Array.size(); ++j) {
-                        Long fieldNameHash = hashCode64Array.get(j);
+                        fieldNameHash = hashCode64Array.get(j);
                         int index = mappingIndex.get(fieldNameHash);
                         AttributeInfo field = mapping.get(fieldNameHash);
-                        Block.IfStmt ifStmt = label.ifStmt(eq(hashCode64, ldc(fieldNameHash)));
-                        genReadFieldValue(ifStmt, index, si, field, jsonReader, object, forLabel, jsonb);
-                        ifStmt.continueStmt(forLabel);
+                        List<JCTree.JCStatement> stmtsIf = genReadFieldValue(field, jsonReaderIdent, index, structInfo, loopLabel, objectIdent, isJsonb);
+                        stmts = stmts.append(defIf(binary(JCTree.Tag.EQ, hashCode64, literal(fieldNameHash)), block(stmtsIf), null));
+                        stmts.append(defContinue(loopLabel));
                     }
-                    label.breakStmt();
+                    stmts.append(defBreak(switchLabel));
                 }
+                cases.append(defCase(getHashCode32Var(literal(fieldNameHash)).getInitializer(), stmts));
             }
-            whileStmt.newLine();
+            switchLabel.body = defSwitch(hashCode32, cases.toList());
+            loopBody.append(switchLabel);
         }
-
-        if (si.smartMatch) {
-            whileStmt.ifStmt(invoke(THIS, "readFieldValueWithLCase", jsonReader, object, hashCode64, features2))
-                    .continueStmt(forLabel);
-            whileStmt.newLine();
+        if (structInfo.smartMatch) {
+            loopBody.append(defIf(method(field(ident(names._this), "readFieldValueWithLCase"), List.of(jsonReaderIdent, objectIdent, ident(hashCode64Var.name), ident(features2Var.name))), block(defContinue(loopLabel)), null));
         }
+        JCTree.JCFieldAccess processExtraField = field(ident(names._this), "processExtra");
+        loopBody.append(exec(method(processExtraField, List.of(jsonReaderIdent, objectIdent))));
+        loopHead.body = block(loopBody.toList());
+        loopLabel.body = loopHead;
+        readObjectBody.append(loopLabel);
 
-        whileStmt.stmt(invoke(null, "processExtra", jsonReader, object));
+        readObjectBody.append(exec(method(field(jsonReaderIdent, "nextIfComma"))));
 
-        mw.newLine();
-        mw.invoke(jsonReader, "nextIfComma");
+        readObjectBody.append(defReturn(objectIdent));
 
-        mw.newLine();
-        mw.ret(object);
+        return defMethod(Flags.PUBLIC, "readObject", objectType, null, List.of(jsonReaderVar, fieldTypeVar, fieldNameVar, featuresVar), null, block(readObjectBody.toList()), null);
     }
 
-    private static void genRead243(
-            StructInfo info,
-            Block.WhileStmt whileStmt,
-            String forLabel,
-            List<AttributeInfo> fields,
-            Op jsonReader,
-            Op object
-    ) {
+    private JCTree.JCVariableDecl getHashCode32Var(JCTree.JCExpression hashCode64) {
+        JCTree.JCBinary usrBinary = binary(JCTree.Tag.USR, hashCode64, literal(TypeTag.INT, 32));
+        JCTree.JCPrimitiveTypeTree intType = type(TypeTag.INT);
+        JCTree.JCTypeCast hashCode32Cast = cast(intType, binary(JCTree.Tag.BITXOR, hashCode64, parens(usrBinary)));
+        return defVar(Flags.PARAMETER, "hashCode32", intType, hashCode32Cast);
+    }
+
+    private List<JCTree.JCStatement> genReadFieldValue(
+            AttributeInfo attributeInfo,
+            JCTree.JCIdent jsonReaderIdent,
+            int i,
+            StructInfo structInfo,
+            JCTree.JCLabeledStatement loopLabel,
+            JCTree.JCIdent objectIdent,
+            boolean isJsonb) {
+        JCTree.JCExpression valueExpr = null;
+        ListBuffer<JCTree.JCStatement> stmts = new ListBuffer<>();
+        String type = attributeInfo.type.toString();
+
+        boolean referenceDetect = structInfo.referenceDetect;
+        if (referenceDetect) {
+            referenceDetect = isReference(type);
+        }
+
+        JCTree.JCFieldAccess fieldReaderField = field(ident(names._this), fieldReader(i));
+        if (referenceDetect) {
+            ListBuffer<JCTree.JCStatement> thenStmts = new ListBuffer<>();
+            JCTree.JCMethodInvocation readReferenceMethod = method(field(jsonReaderIdent, "readReference"));
+            JCTree.JCVariableDecl refVar = defVar(Flags.PARAMETER, "ref", ident("String"), readReferenceMethod);
+            thenStmts.append(refVar);
+            JCTree.JCMethodInvocation addResolveTaskMethod = method(field(fieldReaderField, "addResolveTask"), List.of(jsonReaderIdent, objectIdent, ident(refVar.name)));
+            thenStmts.append(exec(addResolveTaskMethod));
+            thenStmts.append(defContinue(loopLabel));
+            stmts.append(defIf(method(field(jsonReaderIdent, "isReference")), block(thenStmts.toList()), null));
+        }
+
+        String readDirectMethod = getReadDirectMethod(type);
+        if (readDirectMethod != null) {
+            valueExpr = method(field(jsonReaderIdent, readDirectMethod));
+        } else {
+            JCTree.JCExpression fieldValueType = getFieldValueType(type);
+            JCTree.JCVariableDecl fieldValueVar = defVar(Flags.PARAMETER, attributeInfo.name, fieldValueType);
+            stmts.append(fieldValueVar);
+
+            if (type.startsWith("java.util.List<")) {
+                valueExpr = genFieldValueList(type, attributeInfo, jsonReaderIdent, fieldValueVar, loopLabel, stmts, i, referenceDetect, fieldReaderField);
+            } else if (type.startsWith("java.util.Map<java.lang.String,")) {
+                valueExpr = genFieldValueMap(type, attributeInfo, jsonReaderIdent, fieldValueVar, loopLabel, stmts, i, referenceDetect);
+            }
+
+            if (valueExpr == null) {
+                JCTree.JCIdent objectReaderIdent = ident(fieldObjectReader(i));
+                JCTree.JCMethodInvocation getObjectReaderMethod = method(field(fieldReaderField, "getObjectReader"), List.of(jsonReaderIdent));
+                JCTree.JCAssign objectReaderAssign = assign(objectReaderIdent, getObjectReaderMethod);
+                stmts.append(defIf(binary(JCTree.Tag.EQ, objectReaderIdent, defNull()), block(exec(objectReaderAssign)), null));
+                JCTree.JCMethodInvocation objectMethod = method(field(field(ident(names._this), fieldObjectReader(i)), isJsonb ? "readJSONBObject" : "readObject"), List.of(jsonReaderIdent, field(fieldReaderField, "fieldType"), literal(attributeInfo.name), literal(TypeTag.LONG, 0L)));
+                stmts.append(exec(assign(ident(fieldValueVar.name), cast(fieldValueType, objectMethod))));
+                valueExpr = ident(fieldValueVar.name);
+            }
+        }
+
+        if (attributeInfo.setMethod != null) {
+            stmts.append(exec(method(field(objectIdent, attributeInfo.setMethod.getSimpleName().toString()), List.of(valueExpr))));
+        } else if (attributeInfo.field != null) {
+            stmts.append(exec(assign(field(objectIdent, attributeInfo.field.getSimpleName().toString()), valueExpr)));
+        } else {
+            messager.printMessage(Diagnostic.Kind.WARNING, "not implemented yet");
+        }
+        stmts.append(defContinue(loopLabel));
+        return stmts.toList();
+    }
+
+    private List<JCTree.JCStatement> genRead243(
+            java.util.List<AttributeInfo> attributeInfos,
+            JCTree.JCIdent jsonReaderIdent,
+            StructInfo structInfo,
+            JCTree.JCLabeledStatement loopLabel,
+            JCTree.JCIdent objectIdent,
+            boolean isJsonb) {
         IdentityHashMap<AttributeInfo, Integer> readerIndexMap = new IdentityHashMap<>();
-        Map<Integer, List<AttributeInfo>> name0Map = new TreeMap<>();
-        for (int i = 0; i < fields.size(); ++i) {
-            AttributeInfo field = fields.get(i);
+        Map<Integer, java.util.List<AttributeInfo>> name0Map = new TreeMap<>();
+        for (int i = 0; i < attributeInfos.size(); ++i) {
+            AttributeInfo field = attributeInfos.get(i);
             readerIndexMap.put(field, i);
 
             byte[] fieldName = field.name.getBytes(StandardCharsets.UTF_8);
@@ -416,7 +443,7 @@ public class JSONCompiledAnnotationProcessor
 
             int name0 = UNSAFE.getInt(name0Bytes, ARRAY_BYTE_BASE_OFFSET);
 
-            List<AttributeInfo> fieldReaders = name0Map.get(name0);
+            java.util.List<AttributeInfo> fieldReaders = name0Map.get(name0);
             if (fieldReaders == null) {
                 fieldReaders = new ArrayList<>();
                 name0Map.put(name0, fieldReaders);
@@ -434,30 +461,30 @@ public class JSONCompiledAnnotationProcessor
             }
         }
 
-        Op getRawInt = invoke(jsonReader, "getRawInt");
-        Block.SwitchStmt switchStmt = whileStmt.switchStmt(getRawInt, switchKeys);
+        List<JCTree.JCStatement> stmts = List.nil();
+        ListBuffer<JCTree.JCCase> cases = new ListBuffer<>();
 
+        JCTree.JCLabeledStatement switchLabel = label("_switch2", null);
+        ListBuffer<JCTree.JCStatement> switchBody = new ListBuffer<>();
         for (int i = 0; i < labels.length; i++) {
             int name0 = switchKeys[i];
-
-            Block label = switchStmt.label(i);
-
-            List<AttributeInfo> fieldReaders = name0Map.get(name0);
+            java.util.List<AttributeInfo> fieldReaders = name0Map.get(name0);
+            ListBuffer<JCTree.JCStatement> caseStmts = new ListBuffer<>();
             for (int j = 0; j < fieldReaders.size(); j++) {
                 AttributeInfo fieldReader = fieldReaders.get(j);
                 int fieldReaderIndex = readerIndexMap.get(fieldReader);
                 byte[] fieldName = fieldReader.name.getBytes(StandardCharsets.UTF_8);
                 int fieldNameLength = fieldName.length;
-                Opcodes.Op nextIfMatch;
+                JCTree.JCMethodInvocation nextIfMethod;
                 switch (fieldNameLength) {
                     case 2:
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match2");
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match2"));
                         break;
                     case 3:
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match3");
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match3"));
                         break;
                     case 4:
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match4", ldc(fieldName[3]));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match4"), List.of(cast(type(TypeTag.BYTE), literal(TypeTag.INT, fieldName[3]))));
                         break;
                     case 5: {
                         byte[] bytes4 = new byte[4];
@@ -466,7 +493,7 @@ public class JSONCompiledAnnotationProcessor
                         bytes4[2] = '"';
                         bytes4[3] = ':';
                         int name1 = UNSAFE.getInt(bytes4, ARRAY_BYTE_BASE_OFFSET);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match5", ldc(name1));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match5"), List.of(literal(name1)));
                         break;
                     }
                     case 6: {
@@ -476,17 +503,17 @@ public class JSONCompiledAnnotationProcessor
                         bytes4[2] = fieldName[5];
                         bytes4[3] = '"';
                         int name1 = UNSAFE.getInt(bytes4, ARRAY_BYTE_BASE_OFFSET);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match6", ldc(name1));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match6"), List.of(literal(name1)));
                         break;
                     }
                     case 7: {
                         int name1 = UNSAFE.getInt(fieldName, ARRAY_BYTE_BASE_OFFSET + 3);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match7", ldc(name1));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match7"), List.of(literal(name1)));
                         break;
                     }
                     case 8: {
                         int name1 = UNSAFE.getInt(fieldName, ARRAY_BYTE_BASE_OFFSET + 3);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match8", ldc(name1), ldc(fieldName[7]));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match8"), List.of(literal(name1), cast(type(TypeTag.BYTE), literal(TypeTag.INT, fieldName[7]))));
                         break;
                     }
                     case 9: {
@@ -495,7 +522,7 @@ public class JSONCompiledAnnotationProcessor
                         bytes8[6] = '"';
                         bytes8[7] = ':';
                         long name1 = UNSAFE.getLong(bytes8, ARRAY_BYTE_BASE_OFFSET);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match9", ldc(name1));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match9"), List.of(literal(name1)));
                         break;
                     }
                     case 10: {
@@ -503,17 +530,17 @@ public class JSONCompiledAnnotationProcessor
                         System.arraycopy(fieldName, 3, bytes8, 0, 7);
                         bytes8[7] = '"';
                         long name1 = UNSAFE.getLong(bytes8, ARRAY_BYTE_BASE_OFFSET);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match10", ldc(name1));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match10"), List.of(literal(name1)));
                         break;
                     }
                     case 11: {
                         long name1 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 3);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match11", ldc(name1));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match11"), List.of(literal(name1)));
                         break;
                     }
                     case 12: {
                         long name1 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 3);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match12", ldc(name1), ldc(fieldName[11]));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match12"), List.of(literal(name1), cast(type(TypeTag.BYTE), literal(TypeTag.INT, fieldName[11]))));
                         break;
                     }
                     case 13: {
@@ -524,7 +551,7 @@ public class JSONCompiledAnnotationProcessor
                         bytes4[2] = '"';
                         bytes4[3] = ':';
                         int name2 = UNSAFE.getInt(bytes4, ARRAY_BYTE_BASE_OFFSET);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match13", ldc(name1), ldc(name2));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match13"), List.of(literal(name1), literal(name2)));
                         break;
                     }
                     case 14: {
@@ -535,19 +562,19 @@ public class JSONCompiledAnnotationProcessor
                         bytes4[2] = fieldName[13];
                         bytes4[3] = '"';
                         int name2 = UNSAFE.getInt(bytes4, ARRAY_BYTE_BASE_OFFSET);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match14", ldc(name1), ldc(name2));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match14"), List.of(literal(name1), literal(name2)));
                         break;
                     }
                     case 15: {
                         long name1 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 3);
                         int name2 = UNSAFE.getInt(fieldName, ARRAY_BYTE_BASE_OFFSET + 11);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match15", ldc(name1), ldc(name2));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match15"), List.of(literal(name1), literal(name2)));
                         break;
                     }
                     case 16: {
                         long name1 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 3);
                         int name2 = UNSAFE.getInt(fieldName, ARRAY_BYTE_BASE_OFFSET + 11);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match16", ldc(name1), ldc(name2), ldc(fieldName[15]));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match16"), List.of(literal(name1), literal(name2), cast(type(TypeTag.BYTE), literal(TypeTag.INT, fieldName[15]))));
                         break;
                     }
                     case 17: {
@@ -557,7 +584,7 @@ public class JSONCompiledAnnotationProcessor
                         bytes8[6] = '"';
                         bytes8[7] = ':';
                         long name2 = UNSAFE.getLong(bytes8, ARRAY_BYTE_BASE_OFFSET);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match17", ldc(name1), ldc(name2));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match17"), List.of(literal(name1), literal(name2)));
                         break;
                     }
                     case 18: {
@@ -566,19 +593,19 @@ public class JSONCompiledAnnotationProcessor
                         System.arraycopy(fieldName, 11, bytes8, 0, 7);
                         bytes8[7] = '"';
                         long name2 = UNSAFE.getLong(bytes8, ARRAY_BYTE_BASE_OFFSET);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match18", ldc(name1), ldc(name2));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match18"), List.of(literal(name1), literal(name2)));
                         break;
                     }
                     case 19: {
                         long name1 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 3);
                         long name2 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 11);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match19", ldc(name1), ldc(name2));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match19"), List.of(literal(name1), literal(name2)));
                         break;
                     }
                     case 20: {
                         long name1 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 3);
                         long name2 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 11);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match20", ldc(name1), ldc(name2), ldc(fieldName[19]));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match20"), List.of(literal(name1), literal(name2), cast(type(TypeTag.BYTE), literal(TypeTag.INT, fieldName[19]))));
                         break;
                     }
                     case 21: {
@@ -590,7 +617,7 @@ public class JSONCompiledAnnotationProcessor
                         bytes4[2] = '"';
                         bytes4[3] = ':';
                         int name3 = UNSAFE.getInt(bytes4, ARRAY_BYTE_BASE_OFFSET);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match21", ldc(name1), ldc(name2), ldc(name3));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match21"), List.of(literal(name1), literal(name2), literal(name3)));
                         break;
                     }
                     case 22: {
@@ -602,21 +629,21 @@ public class JSONCompiledAnnotationProcessor
                         bytes4[2] = fieldName[21];
                         bytes4[3] = '"';
                         int name3 = UNSAFE.getInt(bytes4, ARRAY_BYTE_BASE_OFFSET);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match22", ldc(name1), ldc(name2), ldc(name3));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match22"), List.of(literal(name1), literal(name2), literal(name3)));
                         break;
                     }
                     case 23: {
                         long name1 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 3);
                         long name2 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 11);
                         int name3 = UNSAFE.getInt(fieldName, ARRAY_BYTE_BASE_OFFSET + 19);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match23", ldc(name1), ldc(name2), ldc(name3));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match23"), List.of(literal(name1), literal(name2), literal(name3)));
                         break;
                     }
                     case 24: {
                         long name1 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 3);
                         long name2 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 11);
                         int name3 = UNSAFE.getInt(fieldName, ARRAY_BYTE_BASE_OFFSET + 19);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match24", ldc(name1), ldc(name2), ldc(name3), ldc(fieldName[23]));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match24"), List.of(literal(name1), literal(name2), literal(name3), cast(type(TypeTag.BYTE), literal(TypeTag.INT, fieldName[23]))));
                         break;
                     }
                     case 25: {
@@ -627,7 +654,7 @@ public class JSONCompiledAnnotationProcessor
                         bytes8[6] = '"';
                         bytes8[7] = ':';
                         long name3 = UNSAFE.getLong(bytes8, ARRAY_BYTE_BASE_OFFSET);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match25", ldc(name1), ldc(name2), ldc(name3));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match25"), List.of(literal(name1), literal(name2), literal(name3)));
                         break;
                     }
                     case 26: {
@@ -637,21 +664,21 @@ public class JSONCompiledAnnotationProcessor
                         System.arraycopy(fieldName, 19, bytes8, 0, 7);
                         bytes8[7] = '"';
                         long name3 = UNSAFE.getLong(bytes8, ARRAY_BYTE_BASE_OFFSET);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match26", ldc(name1), ldc(name2), ldc(name3));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match26"), List.of(literal(name1), literal(name2), literal(name3)));
                         break;
                     }
                     case 27: {
                         long name1 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 3);
                         long name2 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 11);
                         long name3 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 19);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match27", ldc(name1), ldc(name2), ldc(name3));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match27"), List.of(literal(name1), literal(name2), literal(name3)));
                         break;
                     }
                     case 28: {
                         long name1 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 3);
                         long name2 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 11);
                         long name3 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 19);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match28", ldc(name1), ldc(name2), ldc(name3), ldc(fieldName[27]));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match28"), List.of(literal(name1), literal(name2), literal(name3), cast(type(TypeTag.BYTE), literal(TypeTag.INT, fieldName[27]))));
                         break;
                     }
                     case 29: {
@@ -664,7 +691,7 @@ public class JSONCompiledAnnotationProcessor
                         bytes4[2] = '"';
                         bytes4[3] = ':';
                         int name4 = UNSAFE.getInt(bytes4, ARRAY_BYTE_BASE_OFFSET);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match29", ldc(name1), ldc(name2), ldc(name3), ldc(name4));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match29"), List.of(literal(name1), literal(name2), literal(name3), literal(name4)));
                         break;
                     }
                     case 30: {
@@ -677,7 +704,7 @@ public class JSONCompiledAnnotationProcessor
                         bytes4[2] = fieldName[29];
                         bytes4[3] = '"';
                         int name4 = UNSAFE.getInt(bytes4, ARRAY_BYTE_BASE_OFFSET);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match30", ldc(name1), ldc(name2), ldc(name3), ldc(name4));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match30"), List.of(literal(name1), literal(name2), literal(name3), literal(name4)));
                         break;
                     }
                     case 31: {
@@ -685,7 +712,7 @@ public class JSONCompiledAnnotationProcessor
                         long name2 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 11);
                         long name3 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 19);
                         int name4 = UNSAFE.getInt(fieldName, ARRAY_BYTE_BASE_OFFSET + 27);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match31", ldc(name1), ldc(name2), ldc(name3), ldc(name4));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match31"), List.of(literal(name1), literal(name2), literal(name3), literal(name4)));
                         break;
                     }
                     case 32: {
@@ -693,7 +720,7 @@ public class JSONCompiledAnnotationProcessor
                         long name2 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 11);
                         long name3 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 19);
                         int name4 = UNSAFE.getInt(fieldName, ARRAY_BYTE_BASE_OFFSET + 27);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match32", ldc(name1), ldc(name2), ldc(name3), ldc(name4), ldc(fieldName[31]));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match32"), List.of(literal(name1), literal(name2), literal(name3), literal(name4), cast(type(TypeTag.BYTE), literal(TypeTag.INT, fieldName[31]))));
                         break;
                     }
                     case 33: {
@@ -705,7 +732,7 @@ public class JSONCompiledAnnotationProcessor
                         bytes8[6] = '"';
                         bytes8[7] = ':';
                         long name4 = UNSAFE.getLong(bytes8, ARRAY_BYTE_BASE_OFFSET);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match33", ldc(name1), ldc(name2), ldc(name3), ldc(name4));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match33"), List.of(literal(name1), literal(name2), literal(name3), literal(name4)));
                         break;
                     }
                     case 34: {
@@ -716,7 +743,7 @@ public class JSONCompiledAnnotationProcessor
                         System.arraycopy(fieldName, 27, bytes8, 0, 7);
                         bytes8[7] = '"';
                         long name4 = UNSAFE.getLong(bytes8, ARRAY_BYTE_BASE_OFFSET);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match34", ldc(name1), ldc(name2), ldc(name3), ldc(name4));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match34"), List.of(literal(name1), literal(name2), literal(name3), literal(name4)));
                         break;
                     }
                     case 35: {
@@ -724,7 +751,7 @@ public class JSONCompiledAnnotationProcessor
                         long name2 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 11);
                         long name3 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 19);
                         long name4 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 27);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match35", ldc(name1), ldc(name2), ldc(name3), ldc(name4));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match35"), List.of(literal(name1), literal(name2), literal(name3), literal(name4)));
                         break;
                     }
                     case 36: {
@@ -732,7 +759,7 @@ public class JSONCompiledAnnotationProcessor
                         long name2 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 11);
                         long name3 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 19);
                         long name4 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 27);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match36", ldc(name1), ldc(name2), ldc(name3), ldc(name4), ldc(fieldName[35]));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match36"), List.of(literal(name1), literal(name2), literal(name3), literal(name4), cast(type(TypeTag.BYTE), literal(TypeTag.INT, fieldName[35]))));
                         break;
                     }
                     case 37: {
@@ -746,7 +773,7 @@ public class JSONCompiledAnnotationProcessor
                         bytes4[2] = '"';
                         bytes4[3] = ':';
                         int name5 = UNSAFE.getInt(bytes4, ARRAY_BYTE_BASE_OFFSET);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match37", ldc(name1), ldc(name2), ldc(name3), ldc(name4), ldc(name5));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match37"), List.of(literal(name1), literal(name2), literal(name3), literal(name4), literal(name5)));
                         break;
                     }
                     case 38: {
@@ -760,7 +787,7 @@ public class JSONCompiledAnnotationProcessor
                         bytes4[2] = fieldName[37];
                         bytes4[3] = '"';
                         int name5 = UNSAFE.getInt(bytes4, ARRAY_BYTE_BASE_OFFSET);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match38", ldc(name1), ldc(name2), ldc(name3), ldc(name4), ldc(name5));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match38"), List.of(literal(name1), literal(name2), literal(name3), literal(name4), literal(name5)));
                         break;
                     }
                     case 39: {
@@ -769,7 +796,7 @@ public class JSONCompiledAnnotationProcessor
                         long name3 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 19);
                         long name4 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 27);
                         int name5 = UNSAFE.getInt(fieldName, ARRAY_BYTE_BASE_OFFSET + 35);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match39", ldc(name1), ldc(name2), ldc(name3), ldc(name4), ldc(name5));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match39"), List.of(literal(name1), literal(name2), literal(name3), literal(name4), literal(name5)));
                         break;
                     }
                     case 40: {
@@ -778,7 +805,7 @@ public class JSONCompiledAnnotationProcessor
                         long name3 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 19);
                         long name4 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 27);
                         int name5 = UNSAFE.getInt(fieldName, ARRAY_BYTE_BASE_OFFSET + 35);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match40", ldc(name1), ldc(name2), ldc(name3), ldc(name4), ldc(name5), ldc(fieldName[39]));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match40"), List.of(literal(name1), literal(name2), literal(name3), literal(name4), literal(name5), cast(type(TypeTag.BYTE), literal(TypeTag.INT, fieldName[39]))));
                         break;
                     }
                     case 41: {
@@ -792,7 +819,7 @@ public class JSONCompiledAnnotationProcessor
                         bytes8[6] = '"';
                         bytes8[7] = ':';
                         long name5 = UNSAFE.getLong(bytes8, ARRAY_BYTE_BASE_OFFSET);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match41", ldc(name1), ldc(name2), ldc(name3), ldc(name4), ldc(name5));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match41"), List.of(literal(name1), literal(name2), literal(name3), literal(name4), literal(name5)));
                         break;
                     }
                     case 42: {
@@ -804,7 +831,7 @@ public class JSONCompiledAnnotationProcessor
                         System.arraycopy(fieldName, 35, bytes8, 0, 7);
                         bytes8[7] = '"';
                         long name5 = UNSAFE.getLong(bytes8, ARRAY_BYTE_BASE_OFFSET);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match42", ldc(name1), ldc(name2), ldc(name3), ldc(name4), ldc(name5));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match42"), List.of(literal(name1), literal(name2), literal(name3), literal(name4), literal(name5)));
                         break;
                     }
                     case 43: {
@@ -813,258 +840,206 @@ public class JSONCompiledAnnotationProcessor
                         long name3 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 19);
                         long name4 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 27);
                         long name5 = UNSAFE.getLong(fieldName, ARRAY_BYTE_BASE_OFFSET + 35);
-                        nextIfMatch = invoke(jsonReader, "nextIfName4Match43", ldc(name1), ldc(name2), ldc(name3), ldc(name4), ldc(name5));
+                        nextIfMethod = method(field(jsonReaderIdent, "nextIfName4Match43"), List.of(literal(name1), literal(name2), literal(name3), literal(name4), literal(name5)));
                         break;
                     }
                     default:
                         throw new IllegalStateException("fieldNameLength " + fieldNameLength);
                 }
-
-                Block.IfStmt nextIfMatchStmt = label.ifStmt(nextIfMatch);
-                genReadFieldValue(nextIfMatchStmt, fieldReaderIndex, info, fieldReader, jsonReader, object, forLabel, false);
-                nextIfMatchStmt.continueStmt(forLabel);
+                List<JCTree.JCStatement> readFieldValueStmts = genReadFieldValue(fieldReader, jsonReaderIdent, fieldReaderIndex, structInfo, loopLabel, objectIdent, isJsonb);
+                caseStmts.append(defIf(nextIfMethod, block(readFieldValueStmts), null));
             }
-            label.breakStmt();
+            caseStmts.append(defBreak(switchLabel));
+            cases.append(defCase(literal(name0), caseStmts.toList()));
         }
+        switchLabel.body = defSwitch(method(field(jsonReaderIdent, "getRawInt")), cases.toList());
+        stmts = stmts.append(switchLabel);
+        return stmts;
     }
 
-    static void genReadFieldValue(
-            Block mw,
-            int i,
-            StructInfo info,
-            AttributeInfo field,
-            Op jsonReader,
-            Op object,
-            String continueLabel,
-            boolean jsonb
-    ) {
-        Op value = null;
-        String type = field.type.toString();
-
-        boolean referenceDetect = info.referenceDetect;
-        if (referenceDetect) {
-            referenceDetect = CodeGenUtils.isReference(type);
-        }
-
-        if (referenceDetect) {
-            Block.IfStmt isRef = mw.ifStmt(invoke(jsonReader, "isReference"));
-
-            OpName ref = var("ref");
-            isRef.declare(String.class, ref, invoke(jsonReader, "readReference"));
-            isRef.stmt(invoke(var(CodeGenUtils.fieldReader(i)), "addResolveTask", jsonReader, object, ref));
-            isRef.continueStmt(continueLabel);
-            mw.newLine();
-        }
-
-        String readDirectMethod = CodeGenUtils.getReadDirectMethod(type);
-        if (readDirectMethod != null) {
-            value = Opcodes.invoke(jsonReader, readDirectMethod);
-        } else {
-            String fieldName = field.name;
-            OpName fieldValue = var(fieldName);
-            mw.declare(type, fieldValue);
-
-            boolean list = type.startsWith("java.util.List<");
-            boolean mapStr = type.startsWith("java.util.Map<java.lang.String,");
-
-            if (list) {
-                value = genReadFieldValueList(mw, i, jsonReader, type, fieldName, value, fieldValue, referenceDetect);
-            } else if (mapStr) {
-                genReadFieldValueMap(mw, i, jsonReader, type, fieldName, fieldValue, referenceDetect);
-                value = fieldValue;
-            }
-
-            if (value == null) {
-                mw.ifNull(var(CodeGenUtils.fieldObjectReader(i)))
-                        .putField(
-                                CodeGenUtils.fieldObjectReader(i),
-                                invoke(var(CodeGenUtils.fieldReader(i)), "getObjectReader", jsonReader)
-                        );
-                mw.stmt(
-                        assign(
-                                fieldValue,
-                                cast(
-                                        invoke(
-                                                var(fieldObjectReader(i)),
-                                                jsonb ? "readJSONBObject" : "readObject",
-                                                jsonReader,
-                                                getField(var(fieldReader(i)), "fieldType"),
-                                                ldc(field.name),
-                                                ldc(field.readerFeatures)
-                                        ),
-                                        type
-                                )
-                        )
-                );
-            }
-
-            value = fieldValue;
-        }
-
-        if (field.setMethod != null) {
-            mw.invoke(object, field.setMethod.getSimpleName().toString(), value);
-        } else if (field.field != null) {
-            mw.putField(object, field.field.getSimpleName().toString(), value);
-        } else {
-            throw new JSONException("TODO");
-        }
-    }
-
-    private static Op genReadFieldValueList(
-            Block mw,
-            int i,
-            Op jsonReader,
-            String type,
-            String fieldName,
-            Op value,
-            OpName fieldValue,
-            boolean referenceDetect
-    ) {
+    private JCTree.JCExpression genFieldValueList(String type,
+                                                  AttributeInfo attributeInfo,
+                                                  JCTree.JCIdent jsonReaderIdent,
+                                                  JCTree.JCVariableDecl fieldValueVar,
+                                                  JCTree.JCLabeledStatement loopLabel,
+                                                  ListBuffer<JCTree.JCStatement> stmts,
+                                                  int i,
+                                                  boolean referenceDetect,
+                                                  JCTree.JCFieldAccess fieldReaderField) {
         String itemType = type.substring(15, type.length() - 1);
         boolean itemTypeIsClass = itemType.indexOf('<') == -1;
         if (itemTypeIsClass) {
-            Block.IfStmt nextIfNull = mw.ifStmt(invoke(jsonReader, "nextIfNull"));
-            nextIfNull.stmt(assign(fieldValue, ldc(null)));
-            Block nextIfNullElse = nextIfNull.elseStmt();
-            nextIfNullElse.stmt(assign(fieldValue, allocate(ArrayList.class)));
+            JCTree.JCMethodInvocation nextIfNullMethod = method(field(jsonReaderIdent, "nextIfNull"));
+            stmts.append(defIf(nextIfNullMethod,
+                    block(exec(assign(ident(fieldValueVar.name), defNull()))),
+                    block(exec(assign(ident(fieldValueVar.name), newClass(null, null, qualIdent("java.util.ArrayList"), null, null))))));
 
-            String readDirectMethod = CodeGenUtils.getReadDirectMethod(itemType);
-            OpName itemReader = var(CodeGenUtils.fieldItemObjectReader(i));
+            String readDirectMethod = getReadDirectMethod(itemType);
+            JCTree.JCIdent itemReaderIdent = ident(fieldItemObjectReader(i));
             if (readDirectMethod == null) {
-                nextIfNullElse.ifNull(itemReader).stmt(
-                        assign(
-                                itemReader,
-                                invoke(getField(THIS, fieldReader(i)), "getItemObjectReader", jsonReader)
-                        )
-                );
+                JCTree.JCFieldAccess getItemObjectReaderField = field(fieldReaderField, "getItemObjectReader");
+                JCTree.JCExpressionStatement getItemObjectReaderExec = exec(assign(itemReaderIdent, method(getItemObjectReaderField, List.of(jsonReaderIdent))));
+                stmts.append(defIf(binary(JCTree.Tag.EQ, itemReaderIdent, defNull()), block(getItemObjectReaderExec), null));
             }
 
             if (referenceDetect) {
-                referenceDetect = CodeGenUtils.isReference(itemType);
+                referenceDetect = isReference(itemType);
             }
 
-            OpName for_i;
-            if ("i".equals(fieldName)) {
-                for_i = var("j");
+            JCTree.JCVariableDecl for_iVar;
+            if ("i".equals(attributeInfo.name)) {
+                for_iVar = defVar(Flags.PARAMETER, "j", type(TypeTag.INT), literal(0));
             } else {
-                for_i = var("i");
+                for_iVar = defVar(Flags.PARAMETER, "i", type(TypeTag.INT), literal(0));
             }
-            Block.IfStmt nextIfMatch = nextIfNullElse.ifStmt(invoke(jsonReader, "nextIfArrayStart"));
-            Block whileStmt;
-            if (referenceDetect) {
-                whileStmt = nextIfMatch.forStmt(int.class, assign(for_i, ldc(0)), not(invoke(jsonReader, "nextIfArrayEnd")), increment(for_i));
-            } else {
-                whileStmt = nextIfMatch.whileStmtStmt(not(invoke(jsonReader, "nextIfArrayEnd")));
-            }
-
-            Op item;
+            JCTree.JCMethodInvocation nextIfArrayStartMethod = method(field(jsonReaderIdent, "nextIfArrayStart"));
+            JCTree.JCMethodInvocation nextIfArrayEndMethod = method(field(jsonReaderIdent, "nextIfArrayEnd"));
+            ListBuffer<JCTree.JCStatement> whileStmts = new ListBuffer<>();
+            JCTree.JCExpression item;
             if (readDirectMethod != null) {
-                item = invoke(jsonReader, readDirectMethod);
+                item = method(field(jsonReaderIdent, readDirectMethod));
             } else {
-                item = cast(invoke(itemReader, "readObject", jsonReader, ldc(null), ldc(null), ldc(0L)), itemType);
+                item = cast(qualIdent(itemType), method(field(itemReaderIdent, "readObject"), List.of(jsonReaderIdent, defNull(), defNull(), literal(TypeTag.LONG, 0L))));
             }
 
             if (referenceDetect) {
-                OpName listItem = var(fieldName + "_item");
-
-                Block.IfStmt isReference = whileStmt.ifStmt(invoke(jsonReader, "isReference"));
-                OpName ref = var("ref");
-                isReference.declare(String.class, ref, invoke(jsonReader, "readReference"));
-                isReference.invoke(jsonReader, "addResolveTask", fieldValue, for_i, invokeStatic(JSONPath.class, "of", ref));
-                isReference.stmt(invoke(fieldValue, "add", ldc(null)));
-                isReference.continueStmt();
-                whileStmt.newLine();
-
-                whileStmt.declare(itemType, listItem, item);
-                item = listItem;
+                JCTree.JCVariableDecl listItemVar = defVar(Flags.PARAMETER, attributeInfo.name + "_item", getFieldValueType(itemType), item);
+                ListBuffer<JCTree.JCStatement> isReferenceStmts = new ListBuffer<>();
+                JCTree.JCMethodInvocation readReferenceMethod = method(field(jsonReaderIdent, "readReference"));
+                JCTree.JCVariableDecl refVar = defVar(Flags.PARAMETER, "ref", ident("String"), readReferenceMethod);
+                isReferenceStmts.append(refVar);
+                JCTree.JCMethodInvocation addResolveTaskMethod = method(field(jsonReaderIdent, "addResolveTask"), List.of(ident(fieldValueVar.name), ident(for_iVar.name), method(field(qualIdent("com.alibaba.fastjson2.JSONPath"), "of"), List.of(ident(refVar.name)))));
+                isReferenceStmts.append(exec(addResolveTaskMethod));
+                isReferenceStmts.append(exec(method(field(ident(fieldValueVar.name), "add"), List.of(defNull()))));
+                isReferenceStmts.append(defContinue(loopLabel));
+                whileStmts.append(defIf(method(field(jsonReaderIdent, "isReference")), block(isReferenceStmts.toList()), null));
+                whileStmts.append(listItemVar);
+                item = ident(listItemVar.name);
             }
 
-            whileStmt.stmt(invoke(fieldValue, "add", item));
+            whileStmts.append(exec(method(field(ident(fieldValueVar.name), "add"), List.of(item))));
 
-            value = fieldValue;
+            ListBuffer<JCTree.JCStatement> condStmts = new ListBuffer<>();
+            if (referenceDetect) {
+                condStmts.append(forLoop(List.of(for_iVar), unary(JCTree.Tag.NOT, nextIfArrayEndMethod), List.of(exec(unary(JCTree.Tag.PREINC, ident(for_iVar.name)))), block(whileStmts.toList())));
+            } else {
+                condStmts.append(whileLoop(unary(JCTree.Tag.NOT, nextIfArrayEndMethod), block(whileStmts.toList())));
+            }
+
+            stmts.append(defIf(nextIfArrayStartMethod, block(condStmts.toList()), null));
+            return ident(fieldValueVar.name);
         }
-        return value;
+        return null;
     }
 
-    private static void genReadFieldValueMap(
-            Block mw, int i,
-            Op jsonReader,
-            String type,
-            String fieldName,
-            OpName fieldValue,
-            boolean referenceDetect
-    ) {
-        // TODO referenceDetect
+    private JCTree.JCExpression genFieldValueMap(String type,
+                                                 AttributeInfo attributeInfo,
+                                                 JCTree.JCIdent jsonReaderIdent,
+                                                 JCTree.JCVariableDecl fieldValueVar,
+                                                 JCTree.JCLabeledStatement loopLabel,
+                                                 ListBuffer<JCTree.JCStatement> stmts,
+                                                 int i,
+                                                 boolean referenceDetect) {
         String itemType = type.substring(31, type.length() - 1);
         boolean itemTypeIsClass = itemType.indexOf('<') == -1;
-//                    if (itemTypeIsClass) {
-//                    }
-        Block.IfStmt nextIfNull = mw.ifStmt(invoke(jsonReader, "nextIfNull"));
-        nextIfNull.stmt(assign(fieldValue, ldc(null)));
-        Block nextIfNullElse = nextIfNull.elseStmt();
-        OpName itemReader = var(CodeGenUtils.fieldItemObjectReader(i));
+        JCTree.JCMethodInvocation nextIfNullMethod = method(field(jsonReaderIdent, "nextIfNull"));
 
-        boolean readDirect = CodeGenUtils.supportReadDirect(itemType);
+        boolean readDirect = supportReadDirect(itemType);
 
+        ListBuffer<JCTree.JCStatement> elseStmts = new ListBuffer<>();
+        JCTree.JCIdent itemReaderIdent = ident(fieldItemObjectReader(i));
         if (!readDirect) {
-            nextIfNullElse.ifNull(itemReader).stmt(
-                    assign(
-                            itemReader,
-                            invoke(jsonReader, "getObjectReader", getStatic(itemType, "class"))
-                    )
-            );
-            nextIfNullElse.newLine();
+            JCTree.JCFieldAccess getObjectReaderField = field(jsonReaderIdent, "getObjectReader");
+            JCTree.JCExpressionStatement getItemObjectReaderExec = exec(assign(itemReaderIdent, method(getObjectReaderField, List.of(field(qualIdent(itemType), names._class)))));
+            elseStmts.append(defIf(binary(JCTree.Tag.EQ, itemReaderIdent, defNull()), block(getItemObjectReaderExec), null));
         }
 
-        nextIfNullElse.stmt(assign(fieldValue, allocate(HashMap.class)));
+        elseStmts.append(exec(assign(ident(fieldValueVar.name), newClass(null, null, qualIdent("java.util.HashMap"), null, null))));
 
-        nextIfNullElse.newLine();
-        nextIfNullElse.invoke(jsonReader, "nextIfObjectStart");
-        Block.WhileStmt readMapWhile = nextIfNullElse.whileStmtStmt(not(invoke(jsonReader, "nextIfObjectEnd")));
+        JCTree.JCMethodInvocation nextIfObjectStartMethod = method(field(jsonReaderIdent, "nextIfObjectStart"));
+        elseStmts.append(exec(nextIfObjectStartMethod));
+
+        JCTree.JCMethodInvocation nextIfObjectEndMethod = method(field(jsonReaderIdent, "nextIfObjectEnd"));
+        ListBuffer<JCTree.JCStatement> whileStmts = new ListBuffer<>();
 
         if (referenceDetect) {
-            referenceDetect = CodeGenUtils.isReference(itemType);
+            referenceDetect = isReference(itemType);
         }
 
-        Op mapEntryValue;
+        JCTree.JCExpression mapEntryValueExpr;
         if (readDirect) {
-            String method = CodeGenUtils.getReadDirectMethod(itemType);
-            mapEntryValue = invoke(jsonReader, method);
+            mapEntryValueExpr = cast(qualIdent(itemType), method(field(jsonReaderIdent, getReadDirectMethod(itemType))));
         } else {
-            mapEntryValue = cast(
-                    invoke(itemReader, "readObject", jsonReader, getStatic(itemType, "class"), ldc(fieldName), var("features")),
-                    itemType
-            );
+            mapEntryValueExpr = cast(qualIdent(itemType), method(field(itemReaderIdent, "readObject"), List.of(jsonReaderIdent, field(qualIdent(itemType), names._class), literal(attributeInfo.name), ident("features"))));
         }
 
-        Op mapEntryKey = invoke(jsonReader, "readFieldName");
+        JCTree.JCExpression mapEntryKeyExpr = method(field(jsonReaderIdent, "readFieldName"));
 
         if (referenceDetect) {
-            OpName mapKey = var(fieldName + "_key");
-            OpName mapValue = var(fieldName + "_value");
+            JCTree.JCVariableDecl mapKey = defVar(Flags.PARAMETER, attributeInfo.name + "_key", ident("String"), mapEntryKeyExpr);
+            whileStmts.append(mapKey);
+            JCTree.JCVariableDecl mapValue = defVar(Flags.PARAMETER, attributeInfo.name + "_value", ident("String"));
+            whileStmts.append(mapValue);
 
-            readMapWhile.declare(String.class, mapKey, mapEntryKey);
-            mapEntryKey = mapKey;
-
-            readMapWhile.newLine();
-            Block.IfStmt isReference = readMapWhile.ifStmt(invoke(jsonReader, "isReference"));
-            OpName ref = var("ref");
-            isReference.declare(String.class, ref, invoke(jsonReader, "readReference"));
-            isReference.invoke(jsonReader, "addResolveTask", fieldValue, mapKey, invokeStatic(JSONPath.class, "of", ref));
-            isReference.continueStmt();
-            readMapWhile.newLine();
-
-            readMapWhile.declare(itemType, mapValue, mapEntryValue);
-            mapEntryValue = mapValue;
+            ListBuffer<JCTree.JCStatement> isReferenceStmts = new ListBuffer<>();
+            JCTree.JCMethodInvocation readReferenceMethod = method(field(jsonReaderIdent, "readReference"));
+            JCTree.JCVariableDecl refVar = defVar(Flags.PARAMETER, "ref", ident("String"), readReferenceMethod);
+            isReferenceStmts.append(refVar);
+            JCTree.JCMethodInvocation addResolveTaskMethod = method(field(jsonReaderIdent, "addResolveTask"), List.of(ident(fieldValueVar.name), ident(mapKey.name), method(field(qualIdent("com.alibaba.fastjson2.JSONPath"), "of"), List.of(ident(refVar.name)))));
+            isReferenceStmts.append(exec(addResolveTaskMethod));
+            whileStmts.append(defIf(method(field(jsonReaderIdent, "isReference")), block(isReferenceStmts.toList()), exec(method(field(ident(fieldValueVar.name), "put"), List.of(mapEntryKeyExpr, mapEntryValueExpr)))));
+        } else {
+            whileStmts.append(exec(method(field(ident(fieldValueVar.name), "put"), List.of(mapEntryKeyExpr, mapEntryValueExpr))));
         }
 
-        readMapWhile.invoke(
-                fieldValue,
-                "put",
-                mapEntryKey,
-                mapEntryValue
-        );
-        nextIfNullElse.newLine();
-        nextIfNullElse.invoke(jsonReader, "nextIfComma");
+        elseStmts.append(whileLoop(unary(JCTree.Tag.NOT, nextIfObjectEndMethod), block(whileStmts.toList())));
+
+        elseStmts.append(exec(method(field(jsonReaderIdent, "nextIfComma"))));
+
+        stmts.append(defIf(nextIfNullMethod,
+                block(exec(assign(ident(fieldValueVar.name), defNull()))),
+                block(elseStmts.toList())));
+
+        return ident(fieldValueVar.name);
+    }
+
+    private void genSource(String fullQualifiedName, StructInfo structInfo, JCTree.JCClassDecl innerClass) {
+        try {
+            JavaFileObject converterFile = processingEnv.getFiler().createSourceFile(fullQualifiedName, structInfo.element);
+            try (Writer writer = converterFile.openWriter()) {
+                int idx = fullQualifiedName.lastIndexOf(".");
+                if (idx != -1) {
+                    String pkgPath = fullQualifiedName.substring(0, idx);
+                    writer.write("package " + pkgPath + ";");
+                    writer.write(System.lineSeparator());
+                }
+                String str = innerClass.toString().replaceFirst("public static class ", "public class ");
+                writer.write(str);
+            } catch (IOException e) {
+                messager.printMessage(Diagnostic.Kind.ERROR, "Failed saving compiled json serialization file " + fullQualifiedName);
+            }
+        } catch (IOException e) {
+            messager.printMessage(Diagnostic.Kind.ERROR, "Failed creating compiled json serialization file " + fullQualifiedName);
+        }
+    }
+
+    private JCTree.JCExpression getFieldValueType(String type) {
+        if (type.indexOf("[") != -1) {
+            return arrayIdent(type);
+        }
+        if (type.indexOf("<") != -1) {
+            return collectionIdent(type);
+        }
+        return qualIdent(type);
+    }
+
+    private String findConverterName(StructInfo structInfo) {
+        int dotIndex = structInfo.binaryName.lastIndexOf('.');
+        String className = structInfo.binaryName.substring(dotIndex + 1);
+        if (dotIndex == -1) {
+            return className + "_FASTJSONReader";
+        }
+        String packageName = structInfo.binaryName.substring(0, dotIndex);
+        return packageName + '.' + className + "_FASTJSONReader";
     }
 }
