+++ codegen/src/main/resources/META-INF.gradle/incremental.annotation.processors
@@ -1,3 +1,2 @@
 com.alibaba.fastjson2.internal.processor.JSONBaseAnnotationProcessor,DYNAMIC
 com.alibaba.fastjson2.internal.processor.JSONCompiledAnnotationProcessor,DYNAMIC
-com.alibaba.fastjson2.internal.processor.JSONCompiledAnnotationProcessor2,DYNAMIC
