+++ codegen/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,3 +1,2 @@
 com.alibaba.fastjson2.internal.processor.JSONBaseAnnotationProcessor
 com.alibaba.fastjson2.internal.processor.JSONCompiledAnnotationProcessor
-com.alibaba.fastjson2.internal.processor.JSONCompiledAnnotationProcessor2
